fix(views): address PR comments and extract widgets
