+++ assets/translations/en.json
@@ -66,6 +66,7 @@
         "edit_shopping_list": "Edit Shopping List",
         "emptySubtitle": "Create one to get started",
         "emptyTitle": "Nothing found",
+        "errorLoadingProducts": "Error loading products",
         "items_count": "{count} items",
         "loadError": "Error loading collections",
         "noProducts": "No products",
@@ -75,19 +76,24 @@
     "common": {
         "add": "Add",
         "and": " and ",
+        "attemptOf": "Attempt {current} of {total}",
         "cancel": "Cancel",
+        "checkConnection": "Please check your connection and try again later.",
         "close": "Close",
         "create": "Create",
         "delete": "Delete",
         "error_loading_data": "Error loading data",
         "from": "From",
+        "maximumRetriesReached": "Maximum retries reached",
         "name": "Name",
         "name_required": "Please enter a name",
         "or": "OR",
         "privacy_notice": "Privacy Notice",
         "save": "Save",
+        "somethingWentWrong": "Something went wrong",
         "terms_of_service": "Terms of Service",
         "to": "To",
+        "tryAgain": "Try Again",
         "unknown": "Unknown",
         "viewAll": "View All"
     },
@@ -134,6 +140,7 @@
         "details": "Product Details",
         "edit": "Edit Product",
         "enterBarcodeManually": "Enter barcode manually",
+        "errorLoading": "Error loading products",
         "expiration_breakdown": "Expiration Breakdown",
         "expirations": "Expirations",
         "expiredDaysAgo": "Expired {days} days ago",
