+++ assets/translations/nb.json
@@ -66,6 +66,7 @@
         "edit_shopping_list": "Rediger handleliste",
         "emptySubtitle": "Opprett en for å komme i gang",
         "emptyTitle": "Ingenting funnet",
+        "errorLoadingProducts": "Feil ved lasting av produkter",
         "items_count": "{count} varer",
         "loadError": "Feil ved lasting av lister",
         "noProducts": "Ingen produkter",
@@ -75,19 +76,24 @@
     "common": {
         "add": "Legg til",
         "and": " og ",
+        "attemptOf": "Forsøk {current} av {total}",
         "cancel": "Avbryt",
+        "checkConnection": "Vennligst sjekk tilkoblingen din og prøv igjen senere.",
         "close": "Lukk",
         "create": "Opprett",
         "delete": "Slett",
         "error_loading_data": "Feil ved lasting av data",
         "from": "Fra",
+        "maximumRetriesReached": "Maksimalt antall forsøk nådd",
         "name": "Navn",
         "name_required": "Vennligst oppgi et navn",
         "or": "ELLER",
         "privacy_notice": "Personvernerklæring",
         "save": "Lagre",
+        "somethingWentWrong": "Noe gikk galt",
         "terms_of_service": "Vilkår for bruk",
         "to": "Til",
+        "tryAgain": "Prøv igjen",
         "unknown": "Ukjent",
         "viewAll": "Se alle"
     },
@@ -134,6 +140,7 @@
         "details": "Produktdetaljer",
         "edit": "Rediger produkt",
         "enterBarcodeManually": "Skriv inn strekkode manuelt",
+        "errorLoading": "Feil ved lasting av produkter",
         "expiration_breakdown": "Utløpsoversikt",
         "expirations": "Utløpsdatoer",
         "expiredDaysAgo": "Utløpt for {days} dager siden",
