+++ lib/main.dart
@@ -87,9 +87,7 @@ void main() async {
 
     if (!isAlreadyInitialized) rethrow;
 
-    logger.i(
-      'Firebase App Check already initialized, skipping initialization.',
-    );
+    logger.i('Firebase App Check already initialized, skipping...');
   } catch (_) {
     rethrow;
   }
@@ -134,6 +132,8 @@ void main() async {
   );
 }
 
+final ValueNotifier<int> globalRetryNotifier = ValueNotifier<int>(0);
+
 class MyApp extends StatelessWidget {
   final RouterConfig<Object> router;
 
@@ -152,6 +152,28 @@ class MyApp extends StatelessWidget {
         darkTheme: AppTheme.darkTheme,
         themeMode: getIt<ThemeNotifier>().themeMode,
         routerConfig: router,
+        builder: (context, child) {
+          return Stack(
+            children: [
+              ?child,
+              Positioned(
+                right: 16,
+                bottom: 16,
+                child: SafeArea(
+                  child: Material(
+                    type: MaterialType.transparency,
+                    child: FloatingActionButton(
+                      onPressed: () {
+                        globalRetryNotifier.value++;
+                      },
+                      child: const Icon(Icons.refresh),
+                    ),
+                  ),
+                ),
+              ),
+            ],
+          );
+        },
       ),
     );
   }
