+++ lib/widgets/common/retry_scaffold.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:easy_localization/easy_localization.dart';
 import 'package:foodsavr/widgets/common/app_refresh_indicator.dart';
 import 'package:foodsavr/widgets/common/error_state_widget.dart';
+import 'package:foodsavr/main.dart';
 
 /// A scaffold that wraps a view with built-in retry logic, pull-to-refresh,
 /// and error state management.
@@ -27,6 +29,9 @@ class RetryScaffold extends StatefulWidget {
   /// Whether to automatically call [onRefresh] during initState.
   final bool fetchOnInit;
 
+  /// Optional error message to display when loading fails.
+  final String? errorMessage;
+
   const RetryScaffold({
     super.key,
     required this.body,
@@ -36,6 +41,7 @@ class RetryScaffold extends StatefulWidget {
     this.floatingActionButton,
     this.isBodyScrollable = true,
     this.fetchOnInit = true,
+    this.errorMessage,
   });
 
   @override
@@ -45,17 +51,32 @@ class RetryScaffold extends StatefulWidget {
 class _RetryScaffoldState extends State<RetryScaffold> {
   int _retryCount = 0;
   bool _isLoading = false;
+  bool _hasLoadedOnce = false;
   Object? _error;
 
   @override
   void initState() {
     super.initState();
+    globalRetryNotifier.addListener(_onGlobalRetry);
     if (widget.fetchOnInit) {
       _isLoading = true;
       // Use addPostFrameCallback to avoid calling setState during build if onRefresh is extremely fast
       WidgetsBinding.instance.addPostFrameCallback((_) {
         _handleRefresh();
       });
+    }
+  }
+
+  @override
+  void dispose() {
+    globalRetryNotifier.removeListener(_onGlobalRetry);
+    super.dispose();
+  }
+
+  Future<void> _onGlobalRetry() async {
+    if (!mounted) return;
+    if (_error != null && !_isLoading) {
+      await _resetAndRetry();
     }
   }
 
@@ -76,6 +97,7 @@ class _RetryScaffoldState extends State<RetryScaffold> {
         setState(() {
           _retryCount = 0;
           _error = null;
+          _hasLoadedOnce = true;
         });
       }
     } catch (e) {
@@ -83,6 +105,7 @@ class _RetryScaffoldState extends State<RetryScaffold> {
         setState(() {
           _retryCount++;
           _error = e;
+          _hasLoadedOnce = true;
         });
       }
     } finally {
@@ -94,11 +117,11 @@ class _RetryScaffoldState extends State<RetryScaffold> {
     }
   }
 
-  void _resetAndRetry() {
+  Future<void> _resetAndRetry() async {
     setState(() {
       _retryCount = 0;
     });
-    _handleRefresh();
+    await _handleRefresh();
   }
 
   @override
@@ -112,7 +135,7 @@ class _RetryScaffoldState extends State<RetryScaffold> {
 
   Widget _buildBody(BuildContext context) {
     // Initial loading state (only show spinner if it's the very first attempt)
-    if (_isLoading && _retryCount == 0 && _error == null) {
+    if (_isLoading && !_hasLoadedOnce) {
       return const Center(child: CircularProgressIndicator());
     }
 
@@ -126,9 +149,10 @@ class _RetryScaffoldState extends State<RetryScaffold> {
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
             children: [
-              const ErrorStateWidget(
-                message: 'Maximum retries reached',
-                details: 'Please check your connection and try again later.',
+              ErrorStateWidget(
+                message:
+                    widget.errorMessage ?? 'common.maximumRetriesReached'.tr(),
+                details: 'common.checkConnection'.tr(),
               ),
               const SizedBox(height: 24),
               ElevatedButton.icon(
@@ -140,7 +164,7 @@ class _RetryScaffoldState extends State<RetryScaffold> {
                         child: CircularProgressIndicator(strokeWidth: 2),
                       )
                     : const Icon(Icons.refresh),
-                label: const Text('Try Again'),
+                label: Text('common.tryAgain'.tr()),
               ),
             ],
           ),
@@ -160,12 +184,18 @@ class _RetryScaffoldState extends State<RetryScaffold> {
               ),
               const SizedBox(height: 16),
               Text(
-                'Something went wrong',
+                widget.errorMessage ?? 'common.somethingWentWrong'.tr(),
                 style: Theme.of(context).textTheme.titleLarge,
+                textAlign: TextAlign.center,
               ),
               const SizedBox(height: 8),
               Text(
-                'Attempt $_retryCount of ${widget.maxRetries}',
+                'common.attemptOf'.tr(
+                  namedArgs: {
+                    'current': _retryCount.toString(),
+                    'total': widget.maxRetries.toString(),
+                  },
+                ),
                 style: Theme.of(context).textTheme.bodyMedium?.copyWith(
                   color: Theme.of(context).colorScheme.onSurfaceVariant,
                 ),
@@ -180,7 +210,7 @@ class _RetryScaffoldState extends State<RetryScaffold> {
                         child: CircularProgressIndicator(strokeWidth: 2),
                       )
                     : const Icon(Icons.refresh),
-                label: const Text('Try Again'),
+                label: Text('common.tryAgain'.tr()),
               ),
             ],
           ),
